requirements.txt
asgiref==3.4.1
attrs==21.4.0
boto3==1.20.40
botocore==1.23.40
certifi==2021.10.8
charset-normalizer==2.0.10
coverage==6.3
et-xmlfile==1.1.0
idna==3.3
iniconfig==1.1.1
jmespath==0.10.0
numpy==1.22.1
openpyxl==3.0.9
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
Django==4.0.1
pandas==1.4.0
django-model-utils==4.2.0
djangorestframework==3.13.1
